Launch example code
